Result of project
You take the video stream from the raspberry pi camera , do all the
image processing(in order to get the silhouette of the person) on the
raspberry itself and then send it to the led board or pc . This is what
you see on the pc. On the led board you would see almost the same thing
depending on the number of leds used.